mm: swap: fix value of SWP_FAST
SWP_FAST was added to optimize swap usage of zram like fast
swap devices. But its value is greater than SWAP_SCANNING
which is used for refcounting swap map access. This results in
swapoff of the device being blocked forever in a loop waiting
for swap map accesses to complete, by checking the value of
SWAP_SCANNING.

Change-Id: Ibdc07e706c59b6f541708cdb0e28a02c15cfaade
Signed-off-by: Vinayak Menon <vinmenon@codeaurora.org>
Signed-off-by: flar2 <asegaert@gmail.com>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>